soil-veg-app/backend/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
pillow
numpy
python-multipart